Move preferences to utils
